Create Assignment2
I think I have some problems with TortoiseGit because I tried to push some files and every time i did it gave me an error that was saying git did not exit cleanly (exit code 128) instead of the Success message. I googled the error but couldn't find a way to resolv it, I reinstalled TortoiseGit but that didn't help.